requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.2=pypi_0
attrs=22.2.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blas=1.0=mkl
blinker=1.5=pypi_0
bottleneck=1.3.5=py38h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38haa95532_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.5=py38h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icu=58.2=ha925a31_3
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.2=pypi_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=py38hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.0=py38haa95532_0
mdurl=0.1.2=pypi_0
mindmap=0.0.65=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
munkres=1.1.4=py_0
numexpr=2.8.4=py38h5b0cc5e_0
numpy=1.23.5=py38h3b20f71_0
numpy-base=1.23.5=py38h4da318b_0
openssl=1.1.1t=h2bbff1b_0
packaging=22.0=py38haa95532_0
pandas=1.5.3=py38hf11a4ad_0
pcre=8.45=hd77b12b_0
pillow=9.4.0=py38hd77b12b_0
pip=23.0.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
ply=3.11=py38_0
protobuf=3.20.3=pypi_0
pyarrow=11.0.0=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.14.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pyrsistent=0.19.3=pypi_0
python=3.8.16=h6244533_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.20.1=pypi_0
pytz=2022.7=py38haa95532_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.2=pypi_0
rich=13.3.2=pypi_0
semver=2.13.0=pypi_0
setuptools=65.6.3=py38haa95532_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.40.1=h2bbff1b_0
streamlit=1.20.0=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pypi_0
tornado=6.2=py38h2bbff1b_0
typing-extensions=4.5.0=pypi_0
tzdata=2022.7=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.15=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.3.1=pypi_0
wheel=0.38.4=py38haa95532_0
wincertstore=0.2=py38haa95532_2
xz=5.2.10=h8cc25b3_1
zipp=3.11.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
